Modifed UiViewerWindow to update properties table when the current widget is rebuilt and updated UiProfile's rebuild function to disconnect property signals and set the current widget to null.
This fixes the issue where rebuilding the current widget would result in the current (test) widget's signals being
disconnected from the property's setter widget. Also, by delaying the construction of test widget instead of doing
so immediately on a rebuild the property and test widget signals are connected in the expected order, preventing
issues with widget and property signals recursively triggering each other.