ResponsiveLabel progress:
Allow elided text when the TextBox can't contain the smallest label.
Prevent sorting on an update when the updated label hasn't been resized to smaller or larger than a sibling label.
Remove the list's removed label and reassign the stored indices in a single loop.
Fixed off-by-one where a label equal to the widget width wouldn't be displayed.