OrderImbalanceIndicatorTableModel progress:
Added a microsec_clock to hide boost's microsec_clock, allowing custom times to be used during testing without
	modifying the implementation.
Updated row updates and removals to use transactions.
Added TableOperationLog to track every operation done to the model as a side effect of setting intervals and
	offsets when testing.
Added a wait_until function to the test file that processes Qt events while a condition is not met. This allows
	the QtPromises within the model to return a result before the test terminates, without having to
	modify the implementation.